requirements.txt
Flask
line-bot-sdk==2.*
waitress
